Move components into their own folders + fix handleChange bug with inputblock creation
