+++ src/Components/MainTableView/DeleteTuple/DeleteTuple.css
