+++ src/Components/MainTableView/InsertTuple/InsertTuple.css
